Factor cookie extraction out for use elsewhere
